fix problems in test. see http://rb.corp.taobao.com/r/57748/
git-svn-id: http://code.taobao.org/svn/tfs@2130 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-tmp-trunk/src/clientv2/fsname.h
@@ -55,7 +55,7 @@ namespace tfs
         file_.block_id_ = id;
       }
 
-      inline uint32_t get_block_id() const
+      inline uint64_t get_block_id() const
       {
         return file_.block_id_;
       }
